starlist_access::starlist_size() now returns uint32_t.
Formerly it returned int32_t.
Continuing with implementation of heap algorithm in
class Proximate_Stars_View.

+++ output_views.cpp
@@ -1024,7 +1024,7 @@ void BSC_Details_View::put_char( char c ){
   if( c == keypad_return_char ){
     dismiss_action();
   }else if( c == scroll_up_char ){
-    if( index+1 < starlist_access::starlist_size() ){
+    if( static_cast<uint32_t>(index+1) < starlist_access::starlist_size() ){
       set_index( index+1 );
     }
   }else if( c == scroll_down_char ){
